refactor: 라운드 통계 화면 탭과 TabItem 컴포넌트명 수정 #50
