+++ frontend/src/components/RoundVoteContainer/RoundVoteContainer.styled.ts
@@ -1,0 +1,16 @@
+import { css } from '@emotion/react';
+
+export const tabLayout = css`
+  display: flex;
+  flex-basis: 45%;
+  flex-direction: column;
+  width: 100%;
+
+  transition: all 1s;
+`;
+
+export const tabWrapperStyle = css`
+  display: flex;
+  width: 40%;
+  margin-left: 2.4rem;
+`;
