+++ frontend/src/pages/RoundResultPage/RoundResultPage.tsx
@@ -1,13 +1,13 @@
 import NextRoundButton from '@/components/common/NextRoundButton/NextRoundButton';
 import Content from '@/components/layout/Content/Content';
-import RoundResultTab from '@/components/RoundResultTab/RoundResultTab';
+import RoundVoteContainer from '@/components/RoundVoteContainer/RoundVoteContainer';
 import TopicContainer from '@/components/TopicContainer/TopicContainer';
 
 const RoundResultPage = () => {
   return (
     <Content>
       <TopicContainer />
-      <RoundResultTab />
+      <RoundVoteContainer />
       <NextRoundButton />
     </Content>
   );
